hgjghj

+++ eReg/eRegulation/Thermometer.java
@@ -114,6 +114,6 @@ public class Thermometer
     {
     	// Converts temperature in millidegrees into displayable format							// Either keep true or throw it out from display
     	DecimalFormat temperatureFormat 					= new DecimalFormat("#.0");
-    	return  temperatureFormat.format(this.reading/1000);
+    	return  temperatureFormat.format((float) (this.reading)/1000F);
     }
 }
